// frontend/compassClient/models/CreateDaySheetDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateDaySheetDto
 */
export interface CreateDaySheetDto {
    /**
     * 
     * @type {Date}
     * @memberof CreateDaySheetDto
     */
    date?: Date;
    /**
     * 
     * @type {string}
     * @memberof CreateDaySheetDto
     */
    dayReport?: string;
}

/**
 * Check if a given object implements the CreateDaySheetDto interface.
 */
export function instanceOfCreateDaySheetDto(value: object): boolean {
    return true;
}

export function CreateDaySheetDtoFromJSON(json: any): CreateDaySheetDto {
    return CreateDaySheetDtoFromJSONTyped(json, false);
}

export function CreateDaySheetDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateDaySheetDto {
    if (json == null) {
        return json;
    }
    return {
        
        'date': json['date'] == null ? undefined : (new Date(json['date'])),
        'dayReport': json['day_report'] == null ? undefined : json['day_report'],
    };
}

export function CreateDaySheetDtoToJSON(value?: CreateDaySheetDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'date': value['date'] == null ? undefined : ((value['date']).toISOString()),
        'day_report': value['dayReport'],
    };
}